feat: added more type-tests for reflect
