+++ type-tests/types-reflect.tsx
@@ -1,0 +1,105 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+import React from 'react';
+import { expectType } from 'tsd';
+import { createStore, createEvent } from 'effector';
+import { reflect } from '../src';
+
+// basic reflect
+{
+  const Input: React.FC<{
+    value: string;
+    onChange: (newValue: string) => void;
+    color: 'red';
+  }> = () => null;
+  const $value = createStore<string>('');
+  const changed = createEvent<string>();
+
+  const ReflectedInput = reflect({
+    view: Input,
+    bind: {
+      value: $value,
+      onChange: changed,
+      color: 'red',
+    },
+  });
+
+  expectType<React.FC>(ReflectedInput);
+}
+
+// reflect allows infer handlers for event.prepend
+{
+  const Input: React.FC<{
+    value: string;
+    onChange: (event: { target: { value: string } }) => void;
+  }> = () => null;
+  const $value = createStore<string>('');
+  const changed = createEvent<string>();
+
+  const ReflectedInput = reflect({
+    view: Input,
+    bind: {
+      value: $value,
+      // here typescript should infer types correctly
+      onChange: changed.prepend((e) => {
+        expectType<string>(e.target.value);
+        return e.target.value;
+      }),
+    },
+  });
+
+  expectType<React.FC>(ReflectedInput);
+}
+
+// reflect should not allow wrong props
+{
+  const Input: React.FC<{
+    value: string;
+    onChange: (newValue: string) => void;
+    color: 'red';
+  }> = () => null;
+  const $value = createStore<string>('');
+  const changed = createEvent<string>();
+
+  const ReflectedInput = reflect({
+    view: Input,
+    bind: {
+      value: $value,
+      onChange: changed,
+      // @ts-expect-error
+      color: 'blue',
+    },
+  });
+
+  expectType<React.FC>(ReflectedInput);
+}
+
+// reflect should allow not-to pass required props - as they can be added later in react
+{
+  const Input: React.FC<{
+    value: string;
+    onChange: (newValue: string) => void;
+    color: 'red';
+  }> = () => null;
+  const $value = createStore<string>('');
+  const changed = createEvent<string>();
+
+  const ReflectedInput = reflect({
+    view: Input,
+    bind: {
+      value: $value,
+      onChange: changed,
+    },
+  });
+
+  const App: React.FC = () => {
+    // missing prop must still be required in react
+    // @ts-expect-error
+    return <ReflectedInput />;
+  };
+
+  const AppFixed: React.FC = () => {
+    return <ReflectedInput color="red" />;
+  };
+  expectType<React.FC>(App);
+  expectType<React.FC>(AppFixed);
+}
